リストの管理を切り出し
