+++ frontend/src/features/chat/routes/ChatRoomSettingsPage.tsx
@@ -1,24 +1,19 @@
 import * as React from 'react';
 import * as C from '@chakra-ui/react';
 import { ChatRoomMemberStatus, ChatRoom } from '@prisma/client';
-import { useChangeChatRoomMemberStatus } from 'features/chat/hooks/useChangeChatRoomMemberStatus';
 import { useDeleteChatRoom } from 'features/chat/hooks/useDeleteChatRoom';
 import { useExitChatRoom } from 'features/chat/hooks/useExitChatRoom';
 import {
-  ResponseChatMessage,
   ResponseChatRoomMember,
   ResponseChatRoomMemberStatus,
 } from 'features/chat/types/chat';
 import { useGetApi2 } from 'hooks/api/generics/useGetApi2';
-import { useSocket } from 'hooks/socket/useSocket';
 import * as ReactRouter from 'react-router-dom';
 import { ContentLayout } from 'components/ecosystems/ContentLayout';
-import { ChatRoomMemberActionTimeSetModal } from 'features/chat/components/organisms/ChatRoomMemberActionTimeSetModal';
 import { ChatRoomMemberList } from 'features/chat/components/organisms/ChatRoomMemberList';
 import { SecurityAccordionItem } from 'features/chat/components/organisms/SecurityAccordionItem';
 
 export const ChatRoomSettingsPage: React.FC = React.memo(() => {
-  const socket = useSocket(import.meta.env.VITE_WS_CHAT_URL);
   const { chatRoomId } = ReactRouter.useParams() as { chatRoomId: string };
   const { data: chatLoginUserData } = useGetApi2<ResponseChatRoomMember>(
     `/chat/rooms/${chatRoomId}/members/me`
@@ -27,32 +22,8 @@ export const ChatRoomSettingsPage: React.FC = React.memo(() => {
   const { data: chatRoomData } = useGetApi2<ResponseChatRoomMemberStatus>(
     `/chat/rooms/${chatRoomId}`
   );
-  const { data: chatMembersData, refetch: refetchChatMembers } = useGetApi2<
-    ResponseChatMessage[]
-  >(`/chat/rooms/${chatRoomId}/members`);
-  const { isOpen, onClose, changeChatRoomMemberStatus, setSelectedLimitTime } =
-    useChangeChatRoomMemberStatus(chatRoomId, socket);
-  React.useEffect(() => {
-    const fetchDate = async () => {
-      try {
-        await refetchChatMembers();
-      } catch (err) {
-        console.log(err);
-      }
-    };
-    // webSocket
-    socket.emit('join_room_member', chatRoomId);
-    // webSocketのイベントを受け取る関数を登録
-    socket.on('changeChatRoomMemberStatusSocket', fetchDate);
-
-    return () => {
-      socket.emit('leave_room_member', chatRoomId);
-      socket.off('changeChatRoomMemberStatusSocket', fetchDate);
-    };
-  }, []);
-  const chatLoginUser = chatLoginUserData as ResponseChatRoomMember;
-  const chatMembers = chatMembersData as ResponseChatRoomMember[];
   const { name: chatName, roomStatus } = chatRoomData as ChatRoom;
+  const chatLoginUser = chatLoginUserData as ResponseChatRoomMember;
 
   return (
     <>
@@ -61,9 +32,8 @@ export const ChatRoomSettingsPage: React.FC = React.memo(() => {
           {/* ChatRoomMemberListAccordion */}
           <CustomAccordion title="Chat Members">
             <ChatRoomMemberList
+              chatRoomId={chatRoomId}
               chatLoginUser={chatLoginUser}
-              chatMembers={chatMembers}
-              changeChatRoomMemberStatus={changeChatRoomMemberStatus}
             />
           </CustomAccordion>
           {/* SecurityAccordion */}
@@ -81,14 +51,6 @@ export const ChatRoomSettingsPage: React.FC = React.memo(() => {
         {/* ADMINにはDeleteボタンを表示する */}
         <LeaveButton chatRoomId={chatRoomId} chatLoginUser={chatLoginUser} />
       </ContentLayout>
-      <ChatRoomMemberActionTimeSetModal
-        isOpen={isOpen}
-        onClose={onClose}
-        onClick={(limitTime) => {
-          setSelectedLimitTime(limitTime);
-          onClose();
-        }}
-      />
     </>
   );
 });
